requirements/base.txt
# clinic/Backend/requirements/base.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.2.0
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
phonenumbers==8.13.27
python-dateutil==2.8.2
reportlab==4.0.9  # For PDF generation

# clinic/Backend/requirements/local.txt
-r base.txt
ipdb==0.13.13
django-debug-toolbar==4.2.0
django-extensions==3.2.3
python-dotenv==1.0.0

# clinic/Backend/requirements/production.txt  
-r base.txt
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.40.3
django-redis==5.3.0